+++ contracts/agency/libraries/LiquidityAmounts.sol
@@ -6,24 +6,24 @@ import './FixedPoint96.sol';
 
 // solhint-disable
 library LiquidityAmounts {
-  function toUint128(uint256 x) private pure returns (uint128 y) {
+  function toUint128(uint x) private pure returns (uint128 y) {
     require((y = uint128(x)) == x);
   }
 
   function getLiquidityForAmount0(
     uint160 sqrtRatioAX96,
     uint160 sqrtRatioBX96,
-    uint256 amount0
+    uint amount0
   ) internal pure returns (uint128 liquidity) {
     if (sqrtRatioAX96 > sqrtRatioBX96) (sqrtRatioAX96, sqrtRatioBX96) = (sqrtRatioBX96, sqrtRatioAX96);
-    uint256 intermediate = FullMath.mulDiv(sqrtRatioAX96, sqrtRatioBX96, FixedPoint96.Q96);
+    uint intermediate = FullMath.mulDiv(sqrtRatioAX96, sqrtRatioBX96, FixedPoint96.Q96);
     return toUint128(FullMath.mulDiv(amount0, intermediate, sqrtRatioBX96 - sqrtRatioAX96));
   }
 
   function getLiquidityForAmount1(
     uint160 sqrtRatioAX96,
     uint160 sqrtRatioBX96,
-    uint256 amount1
+    uint amount1
   ) internal pure returns (uint128 liquidity) {
     if (sqrtRatioAX96 > sqrtRatioBX96) (sqrtRatioAX96, sqrtRatioBX96) = (sqrtRatioBX96, sqrtRatioAX96);
     return toUint128(FullMath.mulDiv(amount1, FixedPoint96.Q96, sqrtRatioBX96 - sqrtRatioAX96));
@@ -33,8 +33,8 @@ library LiquidityAmounts {
     uint160 sqrtRatioX96,
     uint160 sqrtRatioAX96,
     uint160 sqrtRatioBX96,
-    uint256 amount0,
-    uint256 amount1
+    uint amount0,
+    uint amount1
   ) internal pure returns (uint128 liquidity) {
     if (sqrtRatioAX96 > sqrtRatioBX96) (sqrtRatioAX96, sqrtRatioBX96) = (sqrtRatioBX96, sqrtRatioAX96);
 
@@ -54,17 +54,17 @@ library LiquidityAmounts {
     uint160 sqrtRatioAX96,
     uint160 sqrtRatioBX96,
     uint128 liquidity
-  ) internal pure returns (uint256 amount0) {
+  ) internal pure returns (uint amount0) {
     if (sqrtRatioAX96 > sqrtRatioBX96) (sqrtRatioAX96, sqrtRatioBX96) = (sqrtRatioBX96, sqrtRatioAX96);
 
-    return FullMath.mulDiv(uint256(liquidity) << FixedPoint96.RESOLUTION, sqrtRatioBX96 - sqrtRatioAX96, sqrtRatioBX96) / sqrtRatioAX96;
+    return FullMath.mulDiv(uint(liquidity) << FixedPoint96.RESOLUTION, sqrtRatioBX96 - sqrtRatioAX96, sqrtRatioBX96) / sqrtRatioAX96;
   }
 
   function getAmount1ForLiquidity(
     uint160 sqrtRatioAX96,
     uint160 sqrtRatioBX96,
     uint128 liquidity
-  ) internal pure returns (uint256 amount1) {
+  ) internal pure returns (uint amount1) {
     if (sqrtRatioAX96 > sqrtRatioBX96) (sqrtRatioAX96, sqrtRatioBX96) = (sqrtRatioBX96, sqrtRatioAX96);
 
     return FullMath.mulDiv(liquidity, sqrtRatioBX96 - sqrtRatioAX96, FixedPoint96.Q96);
@@ -75,7 +75,7 @@ library LiquidityAmounts {
     uint160 sqrtRatioAX96,
     uint160 sqrtRatioBX96,
     uint128 liquidity
-  ) internal pure returns (uint256 amount0, uint256 amount1) {
+  ) internal pure returns (uint amount0, uint amount1) {
     if (sqrtRatioAX96 > sqrtRatioBX96) (sqrtRatioAX96, sqrtRatioBX96) = (sqrtRatioBX96, sqrtRatioAX96);
 
     if (sqrtRatioX96 <= sqrtRatioAX96) {
