+++ contracts/agency/peripherals/keepers/Keep3rKeeperDisputable.sol
@@ -14,8 +14,8 @@ abstract contract Keep3rKeeperDisputable is IKeep3rKeeperDisputable, Keep3rDispu
   function slash(
     address _keeper,
     address _bonded,
-    uint256 _bondAmount,
-    uint256 _unbondAmount
+    uint _bondAmount,
+    uint _unbondAmount
   ) public override onlySlasher {
     if (!disputes[_keeper]) revert NotDisputed();
     _slash(_keeper, _bonded, _bondAmount, _unbondAmount);
@@ -33,8 +33,8 @@ abstract contract Keep3rKeeperDisputable is IKeep3rKeeperDisputable, Keep3rDispu
   function _slash(
     address _keeper,
     address _bonded,
-    uint256 _bondAmount,
-    uint256 _unbondAmount
+    uint _bondAmount,
+    uint _unbondAmount
   ) internal {
     if (_bonded != keep3rV1) {
       try IERC20(_bonded).transfer(governance, _bondAmount + _unbondAmount) returns (bool) {} catch (bytes memory) {}
